+++ src/home2/GeometricFigure.java
@@ -3,7 +3,7 @@ package home2;
 abstract class GeometricFigure {
     int weight;
     abstract double getVolume();
-    public static int formula(int v, int m) {
+    public static double formula(int v, int m) {
         return v / m;
     }
 }
